More code review fixes.
