+++ src/pages/Dashboard/Dashboard.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 
-import TimeEntryFormContainer from './Containers/TimeEntryFormContainer.js';
+import TimeEntryForm from './Components/TimeEntryForm/TimeEntryForm.js';
 import TimeEntryList from './Components/TimeEntryList/TimeEntryList.js';
 import './Dashboard.scss';
 
@@ -9,7 +9,7 @@ class Dashboard extends Component {
     return (
         <div className="dashboard">
             <h1 className="dashboard-header">Start tracking!</h1>
-            <TimeEntryFormContainer />
+            <TimeEntryForm />
             <TimeEntryList />
         </div>
     );
